+++ packages/Webkul/Funder/src/Models/SubmissionProxy.php
@@ -1,0 +1,7 @@
+<?php
+
+namespace Webkul\Funder\Models;
+
+use Konekt\Concord\Proxies\ModelProxy;
+
+class SubmissionProxy extends ModelProxy {}